fix: use section end offset to adapt to the expansion of the code buffer (#245)
## What this PR does / why we need it:

Several methods, such as `trampoline_call`, may use
`CodeSection::maybe_expand_to_ensure_remaining` to expand the code
buffer. After expansion, the start/end address of the instruction
section may change. X64 and AArch64 had not met this problem in the
current test cases, but the RISCV, which is in development, does meet
this problem. So it is good to use the section end offset instead.

Testing locally:
- [x] X64: `jeandle` directory
- [ ] AArch64: `jeandle` directory

Signed-off-by: Guoxiong Li <lgxbslgx@gmail.com>